requirements.txt
twython
Pillow
python-socketio[client]